Crate holochain_conductor_api

Re-exports

pub use config::*;

Modules

config

signal_subscription

Structs

ZomeCall
The data provided across an App interface in order to make a zome call

Enums

AdminRequest
Represents the available conductor functions to call over an Admin interface and will result in a corresponding AdminResponse message being sent back over the interface connection. Enum variants follow a general convention of verb_noun as opposed to the noun_verb of AdminResponse.

AdminResponse
Represents the possible responses to an AdminRequest and follows a general convention of noun_verb as opposed to the verb_noun of AdminRequest.

AppRequest
Represents the available Conductor functions to call over an App interface

AppResponse
Responses to requests received on an App interface

CryptoRequest

ExternalApiWireError
Error type that goes over the websocket wire. This intends to be application developer facing so it should be readable and relevant