Struct holochain_conductor_api::conductor::base::Conductor

pub struct Conductor {
    pub key_loader: KeyLoader,
    pub passphrase_manager: Arc<PassphraseManager>,
    pub hash_config: Option<PwHashConfig>,
    // some fields omitted
}

Main representation of the conductor. Holds a HashMap of Holochain instances referenced by ID. A primary point in this struct is load_config(&mut self, config: &Configuration) -> Result<(), String> which takes a config::Configuration struct and tries to instantiate all configured instances. While doing so it has to load DNA files referenced in the configuration. In order to not bind this code to the assumption that there is a filesystem and also enable easier testing, a DnaLoader ()which is a closure that returns a Dna object for a given path string) has to be injected on creation.

Fields
key_loader: KeyLoaderpassphrase_manager: Arc<PassphraseManager>hash_config: Option<PwHashConfig>
Methods

impl Conductor

pub fn from_config(config: Configuration) -> Self

pub fn add_agent_keystore(&mut self, agent_id: String, keystore: Keystore)

pub fn with_signal_channel(self, signal_tx: Sender<Signal>) -> Self

pub fn p2p_bindings(&self) -> Option<Vec<String>>

pub fn config(&self) -> Configuration

pub fn start_signal_multiplexer(&mut self) -> JoinHandle<()>

Starts a new thread which monitors each instance's signal channel and pushes signals out all interfaces the according instance is part of.

pub fn stop_signal_multiplexer(&self)

pub fn start_all_interfaces(&mut self)

pub fn stop_all_interfaces(&mut self)

pub fn stop_interface_by_id(
&mut self,
id: &String
) -> Result<(), HolochainError>

pub fn start_interface_by_id(&mut self, id: &String) -> Result<(), String>

pub fn start_all_static_servers(&mut self) -> Result<(), String>

pub fn start_instance(
&mut self,
id: &String
) -> Result<(), HolochainInstanceError>

pub fn stop_instance(
&mut self,
id: &String
) -> Result<(), HolochainInstanceError>

pub fn start_all_instances(&mut self) -> Result<(), HolochainInstanceError>

Starts all instances

pub fn start_dpki_instance(&mut self) -> Result<(), HolochainInstanceError>

Starts dpki_happ instances

pub fn stop_all_instances(&mut self) -> Result<(), HolochainInstanceError>

Stops all instances

pub fn instances(&self) -> &InstanceMap

pub fn shutdown(&mut self) -> Result<(), HolochainInstanceError>

Stop and clear all instances

pub fn spawn_network(&mut self) -> Result<SpawnResult, HolochainError>

pub fn boot_from_config(&mut self) -> Result<(), String>

Tries to create all instances configured in the given Configuration object. Calls Configuration::check_consistency() first and clears self.instances. The first time we call this, we also initialize the conductor-wide config for use with all instances

pub fn instantiate_from_config(
&mut self,
id: &String,
maybe_config: Option<&mut Configuration>
) -> Result<Holochain, String>

Creates one specific Holochain instance from a given Configuration, id string and DnaLoader.

pub fn build_conductor_api(
&mut self,
instance_id: String,
config: &Configuration
) -> Result<IoHandler, HolochainError>

pub fn agent_config_to_id(
&mut self,
agent_config: &AgentConfiguration
) -> Result<AgentId, HolochainError>

pub fn check_load_key_for_agent(
&mut self,
agent_id: &String
) -> Result<(), String>

Checks if the key for the given agent can be loaded or was already loaded. Will trigger loading if key is not loaded yet. Meant to be used in conductor executable to first try to load all keys (which will trigger passphrase prompts) before bootstrapping the whole config and have prompts appear in between other initialization output.

pub fn get_keystore_for_agent(
&mut self,
agent_id: &String
) -> Result<Arc<Mutex<Keystore>>, String>

Get reference to keystore for given agent ID. If the key was not loaded (into secure memory) yet, this will use the KeyLoader to do so.

pub fn get_keybundle_for_agent(
&mut self,
agent_id: &String
) -> Result<Arc<Mutex<KeyBundle>>, String>

Get reference to the keybundle stored in the keystore for given agent ID. If the key was not loaded (into secure memory) yet, this will use the KeyLoader to do so.

pub fn load_dna(file: &PathBuf) -> HcResult<Dna>

Default DnaLoader that actually reads files from the filesystem

pub fn dna_dir_path(&self) -> PathBuf

pub fn config_path(&self) -> PathBuf

pub fn instance_storage_dir_path(&self) -> PathBuf

pub fn save_config(&self) -> Result<(), HolochainError>

pub fn save_dna(&self, dna: &Dna) -> Result<PathBuf, HolochainError>

pub fn save_dna_to(
&self,
dna: &Dna,
path: PathBuf
) -> Result<PathBuf, HolochainError>

pub fn using_dpki(&self) -> bool

check for determining if the conductor is using dpki to manage instance keys

pub fn dpki_instance_id(&self) -> Option<String>

returns the instance_id of the dpki app if it is configured

pub fn dpki_init_params(&self) -> Option<String>

returns the init_params for the dpki app if it is configured

pub fn dpki_bootstrap(&mut self) -> Result<(), HolochainError>

bootstraps the dpki app if configured