Module holochain_conductor_api::config

Structs

AgentConfiguration
An agent has a name/ID and is optionally defined by a private key that resides in a file

Bridge
A bridge enables an instance to call zome functions of another instance. It is basically an internal interface.

Configuration
Main conductor configuration struct This is the root of the configuration tree / aggregates all other configuration aspects.

DnaConfiguration
A DNA is represented by a DNA file. A hash can optionally be provided, which could be used to validate that the DNA being installed is the DNA that was intended to be installed.

DpkiConfiguration
Configure which app instance id to treat as the DPKI application handler as well as what parameters to pass it on its initialization

InstanceConfiguration
An instance combines a DNA with an agent. Each instance has its own storage configuration.

InstanceReferenceConfiguration

InterfaceConfiguration
Here, interfaces are user facing and make available zome functions to GUIs, browser based web UIs, local native UIs, other local applications and scripts. We currently have:

LoggerConfiguration
This is a config helper structure used to interface with the holochain logging subcrate. Custom rules/filter can be applied to logging, in fact they are used by default in Holochain to filter the logs from its dependencies.

N3hConfig

SignalConfig
Configure which signals to emit, to reduce unwanted signal volume

UiBundleConfiguration
A UI Bundle is a folder containing static assets which can be served as a UI A hash can optionally be provided, which could be used to validate that the UI being installed is the UI bundle that was intended to be installed.

UiInterfaceConfiguration

Enums

InterfaceDriver

NetworkConfig

StorageConfiguration
This configures the Content Addressable Storage (CAS) that the instance uses to store source chain and DHT shard in. There are two storage implementations in cas_implementations so far:

Functions

default_n3h_mode

default_n3h_log_level

default_n3h_persistence_path

default_persistence_dir

load_configuration
Use this function to load a Configuration from a string.

serialize_configuration