Enum holochain_conductor_api::AdminResponse

pub enum AdminResponse {
    Error(ExternalApiWireError),
    DnaRegistered(DnaHash),
    AppInstalled(InstalledApp),
    AdminInterfacesAdded,
    AgentPubKeyGenerated(AgentPubKey),
    DnasListed(Vec<DnaHash>),
    CellIdsListed(Vec<CellId>),
    ActiveAppsListed(Vec<InstalledAppId>),
    AppInterfaceAttached {
        port: u16,
    },
    AppActivated,
    AppDeactivated,
    StateDumped(String),
    AgentInfoAdded,
    AgentInfoRequested(Vec<AgentInfoSigned>),
}

Represents the possible responses to an AdminRequest and follows a general convention of noun_verb as opposed to the verb_noun of AdminRequest.

Will serialize as an object with any contents of the enum on a key data and the enum variant on a key type, e.g. { type: 'app_interface_attached', data: { port: 4000 } }

Variants

Error(ExternalApiWireError)

Can occur in response to any AdminRequest.

There has been an error during the handling of the request. See ExternalApiWireError for variants.

DnaRegistered(DnaHash)

The successful response to an AdminRequest::RegisterDna

AppInstalled(InstalledApp)

The successful response to an AdminRequest::InstallApp.

The resulting InstalledApp contains the App id, the CellNicks and, most usefully, the new CellIds of the newly installed Dnas. See the InstalledApp docs for details.

AdminInterfacesAdded

The succesful response to an AdminRequest::AddAdminInterfaces.

It means the AdminInterfaces have successfully been added

AgentPubKeyGenerated(AgentPubKey)

The succesful response to an AdminRequest::GenerateAgentPubKey.

Contains a new AgentPubKey generated by the Keystore

DnasListed(Vec<DnaHash>)

The successful response to an AdminRequest::ListDnas.

Contains a list of the hashes of all installed Dnas

CellIdsListed(Vec<CellId>)

The succesful response to an AdminRequest::ListCellIds.

Contains a list of all the Cell ids in the conductor

ActiveAppsListed(Vec<InstalledAppId>)

The succesful response to an AdminRequest::ListActiveApps.

Contains a list of all the active App ids in the conductor

AppInterfaceAttached

The succesful response to an AdminRequest::AttachAppInterface.

AppInterfaceApi successfully attached. Contains the port number that was selected (if not specified) by Holochain for running this App interface

Fields of AppInterfaceAttached

port: u16

Networking port of the new AppInterfaceApi

AppActivated

The succesful response to an AdminRequest::ActivateApp.

It means the App was activated successfully

AppDeactivated

The succesful response to an AdminRequest::DeactivateApp.

It means the App was deactivated successfully.

StateDumped(String)

The succesful response to an AdminRequest::DumpState.

The result contains a string of serialized JSON data which can be deserialized to access the full state dump, and inspect the source chain.

AgentInfoAdded

The succesful response to an AdminRequest::AddAgentInfo.

This means the agent info was successfully added to the peer store.

AgentInfoRequested(Vec<AgentInfoSigned>)

The succesful response to an AdminRequest::RequestAgentInfo.

This is all the agent info that was found for the request.