Module holochain_conductor_api::holochain

holochain_conductor_api is a library for instantiating and using holochain instances that each run a holochain DNA, DHT and source chain.

The struct Holochain wraps everything needed to run such an instance.

Examples

extern crate holochain_conductor_api;
extern crate holochain_core_types;
extern crate holochain_core;
extern crate holochain_net;
extern crate holochain_cas_implementations;
extern crate holochain_dpki;
extern crate holochain_sodium;
extern crate tempfile;
extern crate test_utils;
use holochain_conductor_api::{*, context_builder::ContextBuilder};
use holochain_core::nucleus::ribosome::capabilities::CapabilityRequest;
use holochain_core_types::{
    cas::content::Address,
    agent::AgentId,
    dna::Dna,
    json::JsonString,
    signature::Signature,
};
use holochain_dpki::{key_bundle::KeyBundle, seed::SeedType, SEED_SIZE};
use holochain_sodium::secbuf::SecBuf;
use test_utils;

use std::sync::{Arc, Mutex};
use tempfile::tempdir;

// Instantiate a new holochain instance

// Need to get to something like this:
// let dna = holochain_core_types::dna::from_package_file("mydna.dna.json");

// But for now:
let dna = test_utils::create_arbitrary_test_dna();
let dir = tempdir().unwrap();
let storage_directory_path = dir.path().to_str().unwrap();

// We need to provide a cryptographic key that represents the agent.
// Creating a new random one on the fly:
let mut seed = SecBuf::with_insecure(SEED_SIZE);
seed.randomize();

let keybundle = KeyBundle::new_from_seed_buf(&mut seed).unwrap();

// The keybundle's public part is the agent's address
let agent = AgentId::new("bob", keybundle.get_id());

// The instance needs a conductor API with at least the signing callback:
let conductor_api = interface::ConductorApiBuilder::new()
    .with_agent_signature_callback(Arc::new(Mutex::new(keybundle)))
    .spawn();

// The conductor API, together with the storage and the agent ID
// constitute the instance's context:
let context = ContextBuilder::new()
    .with_agent(agent)
    .with_conductor_api(conductor_api)
    .with_file_storage(storage_directory_path)
    .expect("Tempdir should be accessible")
    .spawn();

let mut hc = Holochain::new(dna,Arc::new(context)).unwrap();

// start up the holochain instance
hc.start().expect("couldn't start the holochain instance");

// call a function in the zome code
hc.call("test_zome", CapabilityRequest::new(Address::from("some_token"), Address::from("caller"), Signature::fake()), "some_fn", "{}");

// get the state
{
    let state = hc.state();

    // do some other stuff with the state here
    // ...
}

// stop the holochain instance
hc.stop().expect("couldn't stop the holochain instance");

Structs

Holochain
contains a Holochain application instance