Struct holochain_conductor_api::interface::ConductorApiBuilder

pub struct ConductorApiBuilder { /* fields omitted */ }

ConductorApiBuilder creates IoHandlers that implement RPCs for exposure through interfaces or bridges. This includes zome function calls as well as admin functionality.

Examples for method names are: {instance_id}/{zome}/{cap}/{func} -> a zome call info/list_instances -> Map of InstanceConfigs, keyed by ID admin/... -> see [with_admin_dna_functions]

Each interface has their own handler, and each may be configured differently. This builder makes it convenient to create handlers with different configurations.

Call any sequence of with_* functions on a ConductorApiBuilder object and finalize with spawn() to retrieve the IoHandler.

Methods

impl ConductorApiBuilder

pub fn new() -> Self

pub fn spawn(self) -> IoHandler

Finish the building and retrieve the populated handler

pub fn with_named_instance_config(
self,
instance_name: String,
instance_config: InstanceConfiguration
) -> Self

Add a InstanceConfig for a custom named instance

pub fn with_instance_configs(
self,
instance_configs: Vec<InstanceConfiguration>
) -> Self

Add a vector of InstanceConfig and regard their ID from the config as name.

pub fn with_instances(self, instances: InstanceMap) -> Self

Add several instances with the names given in the InstanceMap

pub fn with_named_instance(
self,
instance_name: String,
instance: Arc<RwLock<Holochain>>
) -> Self

Add a single instance and register it under the given name

pub fn with_admin_dna_functions(self) -> Self

This adds functions to remotely change any aspect of the conductor config. After any change the conductor's config file gets saved. It is guaranteed that the config is either valid after the change or the change does not get applied but instead an error is reported back.

Full list of functions:

admin/dna/install_from_file: Installs a DNA from a given local file. Params:

id: [string] internal handle/name of the newly created DNA config

path: [string] local file path to DNA file

expected_hash: [string] (optional) the hash of this DNA. If this does not match the actual hash, installation will fail.

admin/dna/uninstall Uninstalls a DNA from the conductor config. Recursively also removes (and stops) all instances this DNA is used in. Params:

id: [string] handle of the DNA to be deleted.

admin/dna/list Returns an array of all configured DNAs.

admin/instance/add Creates a new instance and adds it to the config. Does not start the instance nor add it to an interface (see admin/instance/start and admin/interface/add_instance). Params:

id: [string] Name for the new instance

agent_id: [string] Agent to run this instance with

dna_id: [string] DNA to run in this instance

admin/instance/remove Removes an instance. Also remove its any uses of it in interfaces.

id: [string] Which instance to remove?

admin/instance/start Starts a stopped instance or reports an error if the given instance is running already Params:

id: [string] Which instance to start?

admin/instance/stop Stops a running instance or reports an error if the given instance is not running. Params:

id: [string] Which instance to stop?

admin/instance/list Returns an array of all instances that are configured.

admin/instance/running Returns an array of all instances that are running.

admin/interface/add Adds a new DNA / zome / conductor interface (that provides access to zome functions of selected instances and conductor functions, depending on the interfaces config). This also automatically starts the interface. Different from instances, there are no stopped interfaces - every interface that is configured is also active. Params:

id: [string] ID for the new interface

admin: bool Grant access to (these) admin functions?

type: [string] Either "websocket" or "http"

port: [number] Port to bind the server to.

admin/interface/remove Remove an interface from config. This automatically stops the interface as well. Params:

id: [string] Which interface to stop?

admin/interface/add_instance Make a given DNA instance available via a given interface. This restarts the given interface in order to have the change take effect. Params:

interface_id: Which interface to add the instance to?

instance_id: Which instance to add?

admin/interface/remove_instance Remove an instance from a given interface. This restarts the given interface in order to have the change take effect. Params:

interface_id: Which interface to remove the instance from?

instance_id: Which instance to remove?

admin/interface/list Returns an array of all DNA/zome interfaces.

admin/agent/add Add an agent to the conductor configuration that can be used with instances. Params:

id: Handle of this agent configuration as used in the config / other function calls

name: Nickname of this agent configuration

public_address: Public part of this agents key. Has to match the private key in the given key file.

keystore_file: Local path to the file that holds this agent configuration's private key

admin/agent/remove Remove an agent from the conductor config. Params:

id: Which agent to remove?

admin/agent/list Returns an array of all configured agents.

admin/bridge/add Add a bridge between two instances to enable the caller to call the callee's zome functions. Params:

caller_id: ID of the instance that will be able to call into the other instance

callee_id: ID of the instance which's zome functions can be called

handle: Name that the caller uses to reference this bridge and therefore the other instance.

admin/bridge/remove Remove a bridge Params:

caller_id: ID of the instance that can call into the other instance

callee_id: ID of the instance which's zome functions can be called

admin/bridge/list Returns an array of all bridges.

pub fn with_admin_ui_functions(self) -> Self

Adds a further set of functions to the Conductor RPC for managing static UI bundles and HTTP interfaces to these. This adds the following RPC endpoints:

admin/ui/install Install a UI bundle that can later be hosted by an interface Params:

id ID used to refer to this bundle

root_dir Directory to host on the HTTP server

admin/ui/uninstall Uninstall and remove from the config a UI bundle by ID. This will also stop and remove any ui interfaces that are serving this bundle Params:

id ID of the UI bundle to remove

admin/ui/list List all the currently installed UI bundles

admin/ui_interface/add Add a new UI interface to serve a given bundle on a particular port. This can also optionally specify a dna_interface which this UI should connect to. If a dna_interface is included then the route /_dna_connections.json will be available and to instruct the UI as to where it should connect Params:

id ID used to refer to this ui interface

port Port to host the HTTP server on

bundle UI bundle to serve on this port

dna_interface DNA interface this UI can connect to (Optional)

admin/ui_interface/remove Remove an interface by ID Params:

id ID of the UI interface to remove

admin/ui_interface/list List all the UI interfaces

admin/ui_interface/start Start a UI interface given an ID Params:

id ID of the UI interface to start

admin/ui_interface/stop Stop a UI interface given an ID Params:

id ID of the UI interface to stop

pub fn with_agent_signature_callback(
self,
keybundle: Arc<Mutex<KeyBundle>>
) -> Self

pub fn with_test_admin_functions(self) -> Self

Adds extra functionality for running tests via the RPC interface

test/agent/add Adds a test agent. Test agents do not use the regular keystore and instead use the test_keystore_loader. This makes test running much faster Params: - id String Id to assign to the agent to refer to it in interfaces - name String Nickname for the agent. Can be the same as agent_id Returns: Json object containing the newly created agent address

pub fn with_outsource_signing_callback(
self,
agent_id: AgentId,
signing_service_uri: String
) -> Self

pub fn with_agent_keystore_functions(
self,
keystore: Arc<Mutex<Keystore>>
) -> Self