Struct holochain_conductor_api::keystore::Keystore

pub struct Keystore { /* fields omitted */ }

A type for providing high-level crypto functions and managing secrets securely. Keystore can store an arbitrary number of named secrets such as key pairs and seeds. It can be serialized and deserialized with serde and stores secrets in encrypted KeyBlobs, both in the serialized format as well as in memory, as long as secrets are not used. Once a secret is requested, it gets decrypted and cached in the clear in secure memory ([SecBuf]).

Passphrases for de-/encryption are requested from the PassphraseManager that has to be provided on creation.

Keystore makes sure that all secrets are encrypted with the same passphrase, so although every secrete is stored separately in its own KeyBlob, the whole Keystore has a logical passphrase. If a function that requires a passphrase receives a different passphrase from the PassphraseManager as was used to create this keystore, it will fail.

It provides high-level functions for key/seed derivation such as:

[add_seed_from_seed]

[add_key_from_seed]

[add_signing_key_from_seed]

[add_encrypting_key_from_seed]

and a [sign] function for using stored keys to create signatures.

Methods

impl Keystore

pub fn new(
passphrase_manager: Arc<PassphraseManager>,
hash_config: Option<PwHashConfig>
) -> HcResult<Self>

Create a new keystore. This will query passphrase_manager immediately to set a passphrase for the keystore.

pub fn new_standalone(
passphrase_manager: Arc<PassphraseManager>,
hash_config: Option<PwHashConfig>
) -> HcResult<(Self, Base32)>

Create a new keystore for "standalone" use, i.e. not initialized by a DPKI instance

pub fn new_from_file(
path: PathBuf,
passphrase_manager: Arc<PassphraseManager>,
hash_config: Option<PwHashConfig>
) -> HcResult<Self>

Load a keystore from file. This won't ask for a passphrase until a secret is used via the other functions. Secrets will get loaded to memory instantly but stay encrypted until requested.

pub fn change_passphrase(
&mut self,
old_passphrase: &mut SecBuf,
new_passphrase: &mut SecBuf
) -> HcResult<()>

pub fn save(&self, path: PathBuf) -> HcResult<()>

Serialize the keystore to a file.

pub fn list(&self) -> Vec<String>

return a list of the identifiers stored in the keystore

pub fn add(
&mut self,
dst_id_str: &str,
secret: Arc<Mutex<Secret>>
) -> HcResult<()>

adds a secret to the keystore

pub fn add_random_seed(&mut self, dst_id_str: &str, size: usize) -> HcResult<()>

adds a random root seed into the keystore

pub fn get(&mut self, src_id_str: &str) -> HcResult<Arc<Mutex<Secret>>>

gets a secret from the keystore

pub fn add_seed_from_seed(
&mut self,
src_id_str: &str,
dst_id_str: &str,
context: &SeedContext,
index: u64
) -> HcResult<()>

adds a derived seed into the keystore

pub fn add_key_from_seed(
&mut self,
src_id_str: &str,
dst_id_str: &str,
key_type: KeyType
) -> HcResult<Base32>

adds a keypair into the keystore based on a seed already in the keystore returns the public key

pub fn add_signing_key_from_seed(
&mut self,
src_id_str: &str,
dst_id_str: &str
) -> HcResult<Base32>

adds a signing keypair into the keystore based on a seed already in the keystore returns the public key

pub fn add_encrypting_key_from_seed(
&mut self,
src_id_str: &str,
dst_id_str: &str
) -> HcResult<Base32>

adds an encrypting keypair into the keystore based on a seed already in the keystore returns the public key

pub fn add_keybundle_from_seed(
&mut self,
src_id_str: &str,
dst_id_prefix_str: &str
) -> HcResult<(Base32, Base32)>

adds a keybundle into the keystore based on a seed already in the keystore by adding two keypair secrets (signing and encrypting) under the named prefix returns the public keys of the secrets

pub fn add_keybundle(
&mut self,
dst_id_prefix_str: &str,
keybundle: &mut KeyBundle
) -> HcResult<()>

adds a keybundle into the keystore based on an actual keybundle object by adding two keypair secrets (signing and encrypting) under the named prefix

pub fn get_keybundle(&mut self, src_id_prefix_str: &str) -> HcResult<KeyBundle>

adds a keybundle into the keystore based on a seed already in the keystore by adding two keypair secrets (signing and encrypting) under the named prefix returns the public keys of the secrets

pub fn sign(&mut self, src_id_str: &str, data: String) -> HcResult<Signature>

signs some data using a keypair in the keystore returns the signature